% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callback.R
\name{ccallback}
\alias{ccallback}
\alias{callback}
\alias{dyncallback}
\title{Dynamic wrapping of R functions as C callbacks}
\usage{
ccallback(signature, fun, envir = new.env())
}
\arguments{
\item{signature}{character string specifying the
\link[=call-signature]{call signature} of the C function callback type.}

\item{fun}{R function to be wrapped as a C function pointer.}

\item{envir}{the environment in which to evaluate the call to \code{fun}.}
}
\value{
An external pointer to a synthetically generated C function.
}
\description{
Function to wrap R functions as C function pointers.
}
\details{
Callbacks are user-defined functions that are registered in a foreign library
and that are executed at a later time from within that library.
Examples include user-interface event handlers that are registered in GUI
toolkits, and, comparison functions for custom data types to be passed to
generic sort algorithm.

The function \code{ccallback()} wraps an R function \code{fun} as a C function pointer
and returns an external pointer.
The foreign C function type of the wrapped R function is specified by a \link[=call-signature]{call signature} given by \code{signature}.

When the C function pointer is called, a global callback handler (implemented
in C) is executed first, that dynamically creates an R call expression to
\code{fun} using the arguments, passed from C and converted to R, according to the
argument types signature within the \link[=call-signature]{call signature}
specified. See \code{\link[=dyncall]{dyncall()}} for details on the format.

Finally, the handler evaluates the R call expression within the environment
given by \code{envir}. On return, the R return value of \code{fun} is coerced to the C
value, according to the return type signature specified in \code{signature} .

Aggregate by-value callback arguments and returns use the same \verb{<Type>}
signature syntax as \code{\link[=dyncall]{dyncall()}}. An aggregate argument is passed to the R
callback as a raw-backed \code{cdata} object with \code{struct} and \code{typeinfo}
attributes. An aggregate return value must be a raw-backed object for the
same aggregate type and size. Type or storage mismatches disable the callback
and emit a warning.

Aggregate callbacks are supported on the implemented 64-bit x86 and ARM64
dyncallback backends. On unsupported backends, creating a callback whose
signature contains \verb{<Type>} fails early.

If an error occurs during the evaluation, the callback will be disabled for
further invocations. (This behaviour might change in the future.)
}
\note{
The call signature \strong{MUST} match the foreign C callback function type,
otherwise an activated callback call from C can lead to a \strong{fatal R, process
crash}.

A small amount of memory is allocated with each wrapper.
A finalizer function that frees the allocated memory is registered at the
external pointer.
If the external callback function pointer is registered in a C library, a
reference should also be held in R as long as the callback can be activated
from a foreign C run-time context,
otherwise the garbage collector might call the finalizer and the next
invocation of the callback could lead to a \strong{fatal R process crash} as well.
}
\examples{
# Create a function, wrap it to a callback and call it via dyncall:
f <- function(x, y) x + y
cb <- ccallback("ii)i", f)
r <- dyncall(cb, "ii)i", 20, 3)

# Sort vectors directly via 'qsort' C library function using an R callback:
dynbind(c("msvcrt","c","c.so.6"), "qsort(piip)v;")
cb <- ccallback("pp)i", function(px, py) {
    x <- unpack(px, 0, "d")
    y <- unpack(py, 0, "d")
    if (x >  y) return(1) else if (x == y) return(0) else return(-1)
})
x <- rnorm(100)
qsort(x, length(x), 8, cb)
x

}
\references{
Adler, D. (2012) "Foreign Library Interface", \emph{The R Journal},
\strong{4(1)} , 30--40, June 2012.
\url{https://journal.r-project.org/articles/RJ-2012-004/}

Adler, D., Philipp, T. (2008) \emph{DynCall Project}.
\url{https://dyncall.org}
}
\seealso{
See \link[=call-signature]{call signature} for details on call signatures,
\code{\link[=reg.finalizer]{reg.finalizer()}} for details on finalizers.
}
\keyword{interface}
\keyword{programming}
