\name{rdyncall-internal}
\alias{rdyncall-internal}
\alias{getTypeInfo}
\title{rdyncall internal}
\description{Internal functions and objects of rdyncall Package}
\usage{
getTypeInfo(typeName, envir=parent.frame())
}
\arguments{
\item{typeName}{A character string naming a C structure/union type information R object.}
\item{envir}{Specifies an environment.}
}
\keyword{programming}
\keyword{interface}

