% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/struct.R
\name{cstruct}
\alias{cstruct}
\alias{cunion}
\alias{as.ctype}
\alias{cdata}
\alias{$.struct}
\alias{$<-.struct}
\alias{print.struct}
\alias{print.ctype}
\title{Allocation and handling of foreign C aggregate data types}
\usage{
cstruct(sigs, envir = parent.frame())

cunion(sigs, envir = parent.frame())

as.ctype(x, type)

cdata(type)

\method{$}{struct}(x, index)

\method{$}{struct}(x, index) <- value

\method{print}{struct}(x, indent = 0, ...)

\method{print}{ctype}(x, ...)
}
\arguments{
\item{sigs}{character string that specifies several C struct/union type
\code{signature}s.}

\item{envir}{the environment to install S3 type information object(s).}

\item{x}{external pointer or atomic raw vector of S3 class 'struct'.}

\item{type}{S3 \code{\link[=typeinfo]{typeinfo()}} Object or character string that names the
structure type.}

\item{index}{character string specifying the field name.}

\item{value}{value to be converted according to struct/union field type given
by field index.}

\item{indent}{indentation level for pretty printing structures.}

\item{...}{additional arguments to be passed to \code{\link[base:print]{base::print()}} method.}
}
\description{
Functions for allocation, access and registration of foreign C \code{struct} and
\code{union} data type.
}
\details{
References to foreign C data objects are represented by objects of class
'struct'.

Two reference types are supported:
\itemize{
\item \emph{External pointers} returned by \code{\link[=dyncall]{dyncall()}} using a call signature with a
\emph{typed pointer} return type signature and pointers extracted as a result of
\code{\link[=unpack]{unpack()}} and S3 \code{struct} \code{$}-operators.
\item \emph{Internal objects}, memory-managed by R, are allocated by \code{cdata()}: An
atomic \code{raw} storage object is returned, initialized with length equal to
the byte size of the foreign C data type.
}

In order to access and manipulate the data fields of foreign C aggregate data
objects, the \code{$} and \verb{$<-} S3 operator methods can be used.

S3 objects of class \code{struct} have an attribute \code{struct} set to the name of a
\link{typeinfo} object, which provides the run-time type information of a
particular foreign C type.

The run-time type information for foreign C \code{struct} and \code{union} types need
to be registered once via \code{cstruct} and \code{cunion} functions.
The C data types are specified by \code{sigs}, a signature character string.
The formats for both types are described next:

\strong{Structure type signatures} describe the layout of aggregate \code{struct} C
data types.
Type Signatures are used within the \code{field-types}.
Fixed-size array fields are written as a type signature followed by \verb{[N]},
for example \code{C[4]} for \verb{unsigned char[4]} or \verb{<Point>[2]} for two nested
aggregate values.
\code{field-names} consists of space separated identifier names and should match
the number of fields. Integer bitfields are written as \code{name:width} in the
field name list. Unnamed padding bitfields use \verb{:width}; zero-width alignment
bitfields use \verb{:0}.
Optional layout directives can follow the field names before the final
semicolon.

\if{html}{\out{<div class="sourceCode">}}\preformatted{struct-name \{ field-types \} field-names [layout-directives] ;
}\if{html}{\out{</div>}}

Here is an example of a C \code{struct} type:

\if{html}{\out{<div class="sourceCode">}}\preformatted{struct Rect \{
  signed short x, y;
  unsigned short w, h;
\};
}\if{html}{\out{</div>}}

The corresponding structure type signature string is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"Rect\{ssSS\}x y w h;"
}\if{html}{\out{</div>}}

Bitfields keep the ordinary type signature in \code{field-types} and put the bit
width next to the field name:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"Flags\{IIII\}a:1 b:3 :4 c:8;"
}\if{html}{\out{</div>}}

Packed or manually aligned aggregate layouts can be registered with
\verb{@packed}, \verb{@pack(n)} and \verb{@align(n)} directives, where \code{n} must be a
positive power of two. \verb{@packed} is equivalent to \verb{@pack(1)}, \verb{@pack(n)}
caps member alignment at \code{n}, and \verb{@align(n)} raises the final aggregate
alignment to at least \code{n}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{"Packed\{Cd\}c d @packed;"
"Pack4\{Cd\}c d @pack(4);"
"PackedAligned\{Cd\}c d @packed @align(8);"
}\if{html}{\out{</div>}}

\strong{Union type signatures} describe the components of the \code{union} C
data type.
Type signatures are used within the \code{field-types}.
Fixed-size array fields use the same \verb{[N]} suffix as structure fields.
\code{field-names} consists of space separated identifier names and should match
the number of fields. The same layout directives can follow union field
names.

\if{html}{\out{<div class="sourceCode">}}\preformatted{union-name | field-types \} field-names [layout-directives] ;
}\if{html}{\out{</div>}}

Here is an example of a C \code{union} type:

\if{html}{\out{<div class="sourceCode">}}\preformatted{union Value \{
  int anInt;
  float aFloat;
  struct LongValue aStruct
\};
}\if{html}{\out{</div>}}

The corresponding union type signature string is:

\if{html}{\out{<div class="sourceCode">}}\preformatted{"Value|if<LongValue>\}anInt aFloat aStruct;"
}\if{html}{\out{</div>}}

\code{\link[=as.ctype]{as.ctype()}} can be used to \emph{cast} a foreign C data reference to a different
type.
When using an external pointer reference, this can lead quickly to a
\strong{fatal R process crash} - like in C.
}
\examples{
# Specify the following foreign type:
# struct Rect {
#     short x, y;
#     unsigned short w, h;
# }
cstruct("Rect{ssSS}x y w h;")
r <- cdata(Rect)
print(r)
r$x <- 40
r$y <- 60
r$w <- 10
r$h <- 15
print(r)
str(r)
}
\seealso{
\code{\link[=dyncall]{dyncall()}} for type signatures and \code{\link[=typeinfo]{typeinfo()}} for details on run-time type
information S3 objects.
}
\keyword{interface}
\keyword{programming}
