% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynbind.R
\name{dynbind}
\alias{dynbind}
\title{Binding C library functions via thin call wrappers}
\usage{
dynbind(
  libnames,
  signature,
  envir = parent.frame(),
  callmode = "default",
  pattern = NULL,
  replace = NULL,
  funcptr = FALSE
)
}
\arguments{
\item{libnames}{vector of character strings giving short library names of the
shared library to be loaded. See \code{\link[=dynfind]{dynfind()}} for details.
`}

\item{signature}{character string specifying the \emph{library signature} that
determines the set of foreign function names and types. See details.}

\item{envir}{the environment to use for installation of call wrappers.}

\item{callmode}{character string specifying the calling convention, see
details.}

\item{pattern}{\code{NULL} or regular expression character string applied to
symbolic names.}

\item{replace}{\code{NULL} or replacement character string applied to \code{pattern}
part of symbolic names.}

\item{funcptr}{logical, that indicates whether foreign objects refer to
functions (\code{FALSE}, default) or to function pointer variables
(\code{TRUE} rarely needed).}
}
\value{
The function returns a list with two fields:

\item{libhandle}{External pointer returned by \code{\link[=dynload]{dynload()}}.}
\item{unresolved.symbols}{vector of character strings, the names of
unresolved symbols.}

As a side effect, for each wrapper, \code{\link[=dynbind]{dynbind()}} assigns the
\verb{function-name} to  the corresponding call wrapper function in the
environment given by \code{envir}.

If no shared library is found, an error is reported.
}
\description{
Function to bind several foreign functions of a C library via installation of
thin R call wrappers.
}
\details{
\code{dynbind()} makes a set of C functions available to R through installation of
thin call wrappers. The set of functions, including the symbolic name and
function type, is specified by \code{signature}; a character string that encodes a
library signature:

The \strong{library signature} is a compact plain-text format to specify a
set of function bindings.
It consists of function names and corresponding \link{call signature}.
Function bindings are separated by ";" (semicolon);
white spaces (including tab and new line) are allowed before and after
semicolon.

\if{html}{\out{<div class="sourceCode">}}\preformatted{function-name ( call-signature ; ...
}\if{html}{\out{</div>}}

Here is an example that specifies three function bindings to the OpenGL
library:

\if{html}{\out{<div class="sourceCode">}}\preformatted{`"glAccum(If)v ; glClear(I)v ; glClearColor(ffff)v ;"`
}\if{html}{\out{</div>}}

Symbolic names are resolved using the library specified by \code{libnames}
using \code{\link[=dynfind]{dynfind()}} for loading.
For each function, a thin call wrapper function is created using the
following template:

\if{html}{\out{<div class="sourceCode">}}\preformatted{function(...) .dyncall.<MODE> ( <TARGET>, <SIGNATURE>, ... )
}\if{html}{\out{</div>}}

\verb{<MODE>} is replaced by \code{callmode} argument, see \code{\link[=dyncall]{dyncall()}} for details on
calling conventions.
\verb{<TARGET>} is replaced by the external pointer, resolved by the
\verb{function-name}.
\verb{<SIGNATURE>} is replaced by the call signature string contained in
\code{signature}.

The call wrapper is installed in the environment given by \code{envir}.
The assignment name is obtained from the function signature.
If \code{pattern} and \code{replace} is given, a text replacement is applied to the
name before assignment, useful for basic C name space mangling such as
exchanging the prefix.

As a special case, \code{\link[=dynbind]{dynbind()}} supports binding of pointer-to-function
variables, indicated by setting \code{funcptr} to \code{TRUE}, in which case
\verb{<TARGET>} is replaced with the expression \verb{unpack(<TARGET>,"p",0)} in order
to dereference \verb{<TARGET>} as a pointer-to-function variable at call-time.
}
\examples{
\donttest{
info <- dynbind("R", "R_ShowMessage(Z)v; R_rsort(pi)v;")
R_ShowMessage("hello")
}

}
\seealso{
\code{\link[=dyncall]{dyncall()}} for details on call signatures and calling conventions,
\code{\link[=dynfind]{dynfind()}} for details on short library names,
\code{\link[=unpack]{unpack()}} for details on reading low-level memory (e.g. dereferencing of
(function) pointer variables).
}
\keyword{interface}
\keyword{programming}
