% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdyncall-package.R
\docType{package}
\name{rdyncall-package}
\alias{rdyncall}
\alias{rdyncall-package}
\title{Improved Foreign Function Interface (FFI) and Dynamic Bindings to C Libraries}
\description{
The package provides a cross-platform framework for dynamic binding of C
libraries using a flexible Foreign Function Interface (FFI).
The FFI supports almost all fundamental C types, multiple calling conventions,
symbolic access to foreign C struct/union data types and wrapping of R
functions as C callback function pointers.
Dynamic bindings to shared C libraries are data-driven by cross-platform
DynPort specifications. The current \code{dynport()} implementation supports DCF
\code{.dynport} files and generates real R packages from them.
The package includes a variety of technology demos and OS-specific notes for
installation of shared libraries.
}
\details{
\pkg{rdyncall} offers a stack of interoperability technologies for working
with foreign compiled languages using cross-platform portable abstraction
methods.

For R application development, the package facilitates direct access from R
to the C Application Programming Interface (API) of common libraries.
This enables a new style of development: R applications can use low-level
services of portable C libraries.
System-level code can be implemented in R without leaving the language.
C APIs can be explored from within the R interpreter.
Moving the R code from one platform to the other does not involve
recompilation.
Of course, the run-time libraries need to be installed using a standard
procedure of the target Operating-System Distribution.
See \link{rdyncall-demos} for details on this.

For R core development and research, the package provides an improved Foreign
Function Interface (FFI) that can be used to call arbitrary foreign
pre-compiled C code without the need for additional compilation of wrapper
code.
The back-end library is extendable with new calling conventions (such as
Fortran,Pascal,COM,etc.. - which has not been the focus as of this release,
but might be supported officially in the near future).
Basic type-safety checks for argument passing and framework support for
working with foreign C data types such as pointers, arrays, structs and
wrapping of R functions into first-level C callback function pointers
round up this framework.
}
\section{Overview}{
\itemize{
\item Flexible FFI with support for almost all C types, type-safety checks and
multiple calling conventions. See \code{\link[=dyncall]{dyncall()}}.
\item Loading of shared libraries with \emph{automatic unload management} and using
direct access to OS linker. See \code{\link[=dynload]{dynload()}}.
\item Cross-platform naming and loading of shared libraries. See \code{\link[=dynfind]{dynfind()}}.
\item Binding C library functions via thin call wrappers. See \code{\link[=dynbind]{dynbind()}}.
\item Handling of foreign C pointer, array and struct/union data types. See
\code{\link[=pack]{pack()}} and \code{\link[=cstruct]{cstruct()}}.
\item Dynamic wrapping of R functions as C function pointers to be used in C
callbacks. See \link{ccallback}.
\item Dynamic bindings to standard and common C libraries and APIs (functions,
variables, macro constants, enums, struct and union types). See \code{\link[=dynport]{dynport()}}.
}
}

\section{Getting Starrted}{
Several demos ranging from simple FFI calls to the C standard math library up
to more complex 3D OpenGL/SDL Applications are available.
See \code{demos(package = "rdyncall")} for an overview.
Some demos require shared C libraries to be installed in the system.
Please read \link{rdyncall-demos} for details.
}

\section{Supported Platforms}{
The low-level implementation is mainly based on libraries from the DynCall
Project (\url{https://dyncall.org}).
The library suite is distributed as part of the package source tree.

The dyncall and dyncallback libraries implement generic low-level services
with the help of a small amount of hand-written assembly code and careful
modeling of the target machine's calling sequence for each platform to
support.

The DynCall libraries are tested on Linux, Mac OS X, Windows, BSD derivates
and more exotic platforms such as game consoles and Plan9.
Please see the details in the official DynCall manual for full details of the
back-end.
The R Package has been tested on several major R platforms.
}

\examples{
\dontrun{
dynport(SDL2)
dyn.SDL2::SDL_GetPlatform()
}
}
\references{
Adler, D. (2012) "Foreign Library Interface", \emph{The R Journal},
\strong{4(1)} , 30--40, June 2012.
\url{https://journal.r-project.org/articles/RJ-2012-004/}

Adler, D., Philipp, T. (2008) \emph{DynCall Project}.
\url{https://dyncall.org}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hongyuanjia/rdyncall}
  \item \url{https://dyncall.org}
  \item Report bugs at \url{https://github.com/hongyuanjia/rdyncall/issues}
}

}
\keyword{internal}
